代码清理，示例插件默认不注册 Placeholder API 示例。
