+++ src/main/java/org/js4bukkit/io/file/interfaces/FileManagerInterface.java
@@ -15,9 +15,10 @@ public interface FileManagerInterface<T> {
      * 通过 {@link File} 对象获取特定文件类型对象。
      *
      * @param file {@link File}
+     * @param inputStreamFromResource 是否从资源内获取输入流
      * @return 特定文件类型对象
      */
-    T get(File file);
+    T get(File file, boolean inputStreamFromResource);
 
     /**
      * 通过文件名与路径获取特定文件类型对象。
