+++ src/main/resources/maven-dependencies.yml
