+++ src/main/resources/plugins.yml
